fix: add correct export for core module
