+++ packages/useink/src/core/contracts/index.ts
@@ -2,6 +2,7 @@ export * from './call.ts';
 export * from './decodeCallResult.ts';
 export * from './decodeError.ts';
 export * from './getRegistryError.ts';
+export * from './helpers.ts';
 export * from './toContractAbiMessage.ts';
 export * from './toRegistryErrorDecoded.ts';
 export * from './txPaymentInfo.ts';
