+++ packages/useink/src/core/contracts/txPaymentInfo.ts
@@ -1,9 +1,10 @@
 import {
   AccountId,
-  ContractOptions,
   ContractPromise,
+  LazyContractOptions,
   RuntimeDispatchInfo,
   SignerOptions,
+  toContractOptions,
 } from '..';
 
 export async function txPaymentInfo(
@@ -11,7 +12,7 @@ export async function txPaymentInfo(
   message: string,
   caller: AccountId | string,
   params?: unknown[],
-  options?: ContractOptions,
+  options?: LazyContractOptions,
   signerOptions?: Partial<SignerOptions>,
 ): Promise<RuntimeDispatchInfo | undefined> {
   const tx = contract?.tx?.[message];
@@ -20,8 +21,8 @@ export async function txPaymentInfo(
   try {
     const requiresNoArguments = tx.meta.args.length === 0;
     return await (requiresNoArguments
-      ? tx(options || {})
-      : tx(options || {}, ...(params || []))
+      ? tx(toContractOptions(options))
+      : tx(toContractOptions(options), ...(params || []))
     ).paymentInfo(caller, signerOptions);
   } catch (e: unknown) {
     console.error(e);
