+++ packages/useink/src/react/hooks/contracts/types.ts
@@ -1,11 +1,5 @@
 import { ChainId } from '../../../chains/types.ts';
-import { Abi, ContractOptions, ContractPromise } from '../../../core/index';
-
-export type CallOptions = ContractOptions & {
-  defaultCaller?: boolean;
-};
-
-export type ContractAbi = string | Record<string, unknown> | Abi;
+import { ContractPromise } from '../../../core/index';
 
 export interface ChainContract<T extends ContractPromise = ContractPromise> {
   contract: T;
