+++ packages/useink/src/react/hooks/contracts/useCallSubscription.ts
@@ -1,5 +1,6 @@
+import { LazyCallOptions } from '../../../core/index.ts';
 import { useBlockHeader } from '../substrate/useBlockHeader.ts';
-import { CallOptions, ChainContract } from './types.ts';
+import { ChainContract } from './types.ts';
 import { Call, useCall } from './useCall.ts';
 import { useEffect } from 'react';
 
@@ -7,7 +8,7 @@ export function useCallSubscription<T>(
   chainContract: ChainContract | undefined,
   message: string,
   args = [] as unknown[],
-  options?: CallOptions,
+  options?: LazyCallOptions,
 ): Omit<Call<T>, 'send'> {
   const call = useCall<T>(chainContract, message);
   const blockNumber = useBlockHeader(chainContract?.chainId)?.blockNumber;
