+++ packages/useink/tsup.config.ts
@@ -5,7 +5,7 @@ export default defineConfig({
   format: ['cjs', 'esm'],
   entry: {
     index: 'src/index.ts',
-    core: 'src/index.ts',
+    core: 'src/core/index.ts',
     chains: 'src/chains/index.ts',
     notifications: 'src/notifications/index.ts',
     utils: 'src/utils/index.ts',
